chore(suite): flatten folder structure for AccountsMenu components
