+++ packages/suite/src/components/wallet/AccountsMenu/AnimationWrapper.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { AnimatePresence, motion } from 'framer-motion';
 
-interface Props {
+interface AnimationWrapperProps {
     opened: boolean;
     children?: React.ReactNode;
     onAnimationStart?: () => void;
@@ -9,28 +9,33 @@ interface Props {
     onAnimationComplete?: () => void;
 }
 
-const AnimationWrapper = ({
+export const AnimationWrapper = ({
     opened,
     children,
     onAnimationStart,
     onUpdate,
     onAnimationComplete,
-}: Props) => {
+}: AnimationWrapperProps) => {
     const collapsed = { overflow: 'hidden', height: 0 };
+
+    const animate = {
+        height: 'auto',
+        transitionEnd: { overflow: 'unset' },
+    };
+
+    const transition = { duration: 0.24, ease: 'easeInOut' };
+
     return (
         <AnimatePresence initial={false}>
             {opened && (
                 <motion.div
                     initial={collapsed}
-                    animate={{
-                        height: 'auto',
-                        transitionEnd: { overflow: 'unset' },
-                    }}
+                    animate={animate}
                     exit={collapsed}
                     onAnimationStart={onAnimationStart}
                     onUpdate={onUpdate}
                     onAnimationComplete={onAnimationComplete}
-                    transition={{ duration: 0.24, ease: 'easeInOut' }}
+                    transition={transition}
                 >
                     {children}
                 </motion.div>
@@ -38,5 +43,3 @@ const AnimationWrapper = ({
         </AnimatePresence>
     );
 };
-
-export default AnimationWrapper;
